+++ packages/@expo/cli/src/export/embed/exportEmbedAsync.ts
@@ -110,6 +110,9 @@ export async function createMetroServerAndBundleRequestAsync(
     {
       exp,
       isExporting: true,
+      getMetroBundler() {
+        return server.getBundler().getBundler();
+      },
     }
   );